src/test/kotlin/com/clouddjr/advent2015/Day23Test.kt
package com.clouddjr.advent2015

import com.clouddjr.advent2015.Resources.resourceAsList
import org.junit.jupiter.api.DisplayName
import org.junit.jupiter.api.Nested
import org.junit.jupiter.api.Test
import kotlin.test.assertEquals

@DisplayName("Day 23")
class Day23Test {

    @Nested
    @DisplayName("Part 1")
    inner class Part1 {
        @Test
        fun `Matches actual`() {
            val answer = Day23(resourceAsList("day23.txt")).solvePart1()

            assertEquals(255, answer)
        }
    }

    @Nested
    @DisplayName("Part 2")
    inner class Part2 {
        @Test
        fun `Matches actual`() {
            val answer = Day23(resourceAsList("day23.txt")).solvePart2()

            assertEquals(334, answer)
        }
    }
}